fix(DO-378): mount sledger.yaml into default /ledger directory (#9)
* fix(DO-378): mount sledger.yaml into default /ledger directory

Jira: https://motorefi.atlassian.net/browse/DO-378

* Only kill side car when successful

Jira: https://motorefi.atlassian.net/browse/DO-378

+++ entrypoint.sh
@@ -19,7 +19,7 @@ function run_sledger() {
    result=$?
    
    # after sledger runs, kill the proxy
-   if pgrep cloud_sql_proxy
+   if [ $(pgrep cloud_sql_proxy) ] && [ ${result} -eq 0 ]
    then
       pkill cloud_sql_proxy
    fi
